Chore: Retry failing editor test in CI
