+++ packages/app-mobile/components/FeedbackBanner.test.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { Store } from 'redux';
 import { AppState } from '../utils/types';
 import TestProviderStack from './testing/TestProviderStack';
-import { switchClient, setupDatabase, mockMobilePlatform, mockFetch, waitFor } from '@joplin/lib/testing/test-utils';
+import { switchClient, setupDatabase, mockMobilePlatform, mockFetch } from '@joplin/lib/testing/test-utils';
+import waitFor from '@joplin/lib/testing/waitFor';
 import createMockReduxStore from '../utils/testing/createMockReduxStore';
 import setupGlobalStore from '../utils/testing/setupGlobalStore';
 import { act, fireEvent, render, screen } from '@testing-library/react-native';
