+++ packages/lib/testing/test-utils.ts
@@ -1137,34 +1137,6 @@ export const mockMobilePlatform = (platform: MobilePlatform) => {
 	};
 };
 
-// Waits for callback to not throw. Similar to react-native-testing-library's waitFor, but works better
-// with Joplin's mix of real and fake Jest timers.
-const realSetTimeout = setTimeout;
-export const waitFor = async (callback: ()=> Promise<void>) => {
-	const timeout = 10_000;
-	const startTime = performance.now();
-	let passed = false;
-	let lastError: Error|null = null;
-
-	while (!passed && performance.now() - startTime < timeout) {
-		try {
-			await callback();
-			passed = true;
-			lastError = null;
-		} catch (error) {
-			lastError = error;
-
-			await new Promise<void>(resolve => {
-				realSetTimeout(() => resolve(), 10);
-			});
-		}
-	}
-
-	if (lastError) {
-		throw lastError;
-	}
-};
-
 export const runWithFakeTimers = async (callback: ()=> Promise<void>) => {
 	if (typeof jest === 'undefined') {
 		throw new Error('Fake timers are only supported in jest.');
